Convert 'zfs' pool to a new 'zroot' pool with Root-on-ZFS setup.
Shift all the instructions up to the top of the guide. The FreeBSD installer
should be used for this, however given my system was converted to ZFS only
after being installed, here's how the current config was done. Additionally,
the original 'zfs' raidz pool was migrated live to a 4 disk 'zroot' setup in
a RAID-10 like configuration. Since this was all done on a live system the
FreeBSD installer wasn't used.

While here, remove the encrypted swap that was on the original SSD. I would
like crash dumps when they happen so the cost of encrypted swap don't outweigh
the benefits when it is on a physically safe system.