tftpboot: remove the section due to changing in my requirements
The motivation for this section was two-fold:

One, at the time I was using Foreman as part of a full stack build system at
work and there was benefit to keeping some familiarity with the interactions
of components. The hands on setup was helpful in learning this as compared
to the full foreman-installer configured setup that I used on *el at work.

Two, at the time I was tinkering with Ubuntu desktop OS's more frequently.
As of today I generally just follow the LTS version (still meaning to try
a *BSD on the desktop at some point) but with everything I use plus support
for things like Steam/Proton on the Linux desktop working just find I have
no intent to change this. It's a low cost for spending a few hours on USB
reinstalls every 2 years. With the desire to use UEFI/Secureboot on newer
systems as well that seems easier than really tinkering to vet out fully
UEFI/Secureboot working Linux installs for me. As such, it's time to make
this a bit more simple and leave this in the history until such a time to
justify exploring it again.